ChartProvider

Documentation

draw :: String -> [(String, Int)] -> IO String

Draws a chart of tweets per day and saves it to "tmp{twitterAccountName}.png" file.

drawChart :: String -> String -> [(String, Int)] -> IO ()

Draws a chart of tweets per day and saves it to the given file.