DateParser

Documentation

data TwitterDate

A type class which can represents date returned by the Twitter API.

Constructors

TwitterDate

Fields

day :: Day

weekDay :: WeekDay

month :: Month

year :: Year

time :: Time

zone :: Zone

data Day

Constructors

Day Integer

data WeekDay

Constructors

WeekDay String

data Month

Constructors

Month String

data Year

Constructors

Year Integer

data Hour

Constructors

Hour Int

data Minute

Constructors

Minute Int

data Second

Constructors

Second Int

data Zone

Constructors

Zone String

regularParse :: Parser a -> String -> Either ParseError a

A simple utility function which sets the filePath as an empty string: | Quote from documentation: "The filePath is only used in error messages and may be the empty string."

dayParser :: Parser Day

A simple parser that validates the given input and returns an instance of Day or fails to match if the | input does not conform to requested type. | Note that for simplicity we do not validate correctness in terms of the month - which means that | days like 31 Feb will be consider valid.

weekDayParser :: Parser WeekDay

A simple parser that validates the given input and returns an instance of WeekDay or fails to match if the | input does not conform to requested type.

monthParser :: Parser Month

A simple parser that validates the given input and returns an instance of Month or fails to match if the | input does not conform to requested type.

yearParser :: Parser Year

A simple parser that validates the given input and returns an instance of WeekDay or fails to match if the | input does not conform to requested type.

hourParser :: Parser Hour

A simple parser that validates the given input and returns an instance of Hour or fails to match if the | input does not conform to requested type.

minuteParser :: Parser Minute

A simple parser that validates the given input and returns an instance of Minute or fails to match if the | input does not conform to requested type.

timeParser :: Parser Time

A parser that can parse an input of format dd:mm:ss to an instance of Time.

zoneParser :: Parser Zone

A simple parser that validates the given input and returns an instance of Zone or fails to match if the | input does not conform to requested type.

twitterDateParser :: Parser TwitterDate

A parser that can parse a date like "Thu Oct 27 00:14:27 +0000 2016" to an instance of TwitterDate.

parseDateFromString :: String -> Either ParseError TwitterDate

Tries to parse the given String an instance of TwitterDate. | Notice that thanks to `<* eof` parse is successful only if the whole string meets the predicate, | without it dates like "Thu Oct 27 00:14:27 +0000 201611111" would be cut to "Thu Oct 27 00:14:27 +0000 2016" | without reporting failure

simpleDayString :: TwitterDate -> String

Converts the given TwitterDate to a string representing a single day in year. Example: "29 Dec"